+++ includes/Sonoff.php
@@ -22,14 +22,39 @@
 			return $device;
 		}
 		
-		public function getDevices() {
+		
+		public function getDevices( $orderBy = "position" ) {
 			
 			$devices = [];
-			$file    = fopen( _CSVFILE_, 'r' );
+			
+			$file = fopen( _CSVFILE_, 'r' );
 			while ( ( $line = fgetcsv( $file ) ) !== FALSE ) {
 				$devices[] = $this->createDeviceObject( $line );
+				
+				
 			}
 			fclose( $file );
+			
+			if ( $orderBy == "position" ) {
+				$devicesTmp = [];
+				$update     = FALSE;
+				foreach ( $devices as $device ) {
+					if ( $device->position == "" ) {
+						$device->position = 1;
+						$update           = TRUE;
+					}
+					while ( isset( $devicesTmp[ $device->position ] ) ) {
+						$device->position++;
+					}
+					if ( $update ) {
+						$this->setDeviceValue( $device->id, "position", $device->position );
+					}
+					$devicesTmp[ $device->position ] = $device;
+				}
+				ksort( $devicesTmp );
+				$devices = $devicesTmp;
+				unset( $devicesTmp );
+			}
 			
 			return $devices;
 		}
@@ -47,10 +72,99 @@
 			$device->username = isset( $deviceLine[ 3 ] ) ? $deviceLine[ 3 ] : FALSE;
 			$device->password = isset( $deviceLine[ 4 ] ) ? $deviceLine[ 4 ] : FALSE;
 			$device->img      = isset( $deviceLine[ 5 ] ) ? $deviceLine[ 5 ] : "bulb_1";
+			$device->position = isset( $deviceLine[ 6 ] ) && $deviceLine[ 6 ] != "" ? $deviceLine[ 6 ] : "";
 			
 			return $device;
 		}
 		
+		public function setDeviceValue( $id = NULL, $field = NULL, $value = NULL ) {
+			if ( !isset( $id ) || empty( $id ) ) {
+				return NULL;
+			}
+			$device = NULL;
+			$file   = fopen( _CSVFILE_, 'r' );
+			while ( ( $line = fgetcsv( $file ) ) !== FALSE ) {
+				if ( $line[ 0 ] == $id ) {
+					$device = $this->createDeviceObject( $line );
+					break;
+				}
+			}
+			fclose( $file );
+			$device->$field = $value;
+			$device         = $this->updateDevice( $device );
+			
+			return $device;
+		}
+		
+		public function updateDevice( $device = NULL ) {
+			if ( !isset( $device ) || empty( $device ) || !isset( $device->id ) || empty( $device->id ) ) {
+				return NULL;
+			}
+			$deviceArr[ 0 ] = $device->id;
+			$deviceArr[ 1 ] = implode( "|", isset( $device->names ) && !empty( $device->names ) ? $device->names : [] );
+			$deviceArr[ 2 ] = isset( $device->ip ) && !empty( $device->ip ) ? $device->ip : "";
+			$deviceArr[ 3 ] = isset( $device->username ) && !empty( $device->username ) ? $device->username : "";
+			$deviceArr[ 4 ] = isset( $device->password ) && !empty( $device->password ) ? $device->password : "";
+			$deviceArr[ 5 ] = isset( $device->img ) && !empty( $device->img ) ? $device->img : "";
+			$deviceArr[ 6 ] = isset( $device->position ) && !empty( $device->position ) ? $device->position : "";
+			
+			foreach ( $deviceArr as $key => $field ) {
+				if ( is_array( $field ) ) {
+					foreach ( $field as $subkey => $subfield ) {
+						$deviceArr[ $key ][ $field ][ $subkey ] = trim( $subfield );
+					}
+				} else {
+					
+					$deviceArr[ $key ] = trim( $field );
+				}
+			}
+			
+			$tempfile = @tempnam( _TMPDIR_, "tmp" ); // produce a temporary file name, in the current directory
+			
+			
+			if ( !$input = fopen( _CSVFILE_, 'r' ) ) {
+				die( __( "ERROR_CANNOT_READ_CSV_FILE", "DEVICE_ACTIONS", [ "csvFilePath" => _CSVFILE_ ] ) );
+			}
+			if ( !$output = fopen( $tempfile, 'w' ) ) {
+				die( __( "ERROR_CANNOT_CREATE_TMP_FILE", "DEVICE_ACTIONS", [ "tmpFilePath" => $tempfile ] ) );
+			}
+			
+			while ( ( $data = fgetcsv( $input ) ) !== FALSE ) {
+				if ( $data[ 0 ] == $deviceArr[ 0 ] ) {
+					$data = $deviceArr;
+				}
+				fputcsv( $output, $data );
+			}
+			
+			fclose( $input );
+			fclose( $output );
+			
+			unlink( _CSVFILE_ );
+			rename( $tempfile, _CSVFILE_ );
+			
+			return $this->createDeviceObject( $deviceArr );
+		}
+		
+		public function addDevice( $device = [] ) {
+			die( "not done yet" ); //todo: use this to add device, decide if array or object param
+			if ( !isset( $device ) || empty( $device ) || !isset( $device[ "id" ] ) || empty( $device[ "id" ] ) ) {
+				return NULL;
+			}
+			
+			$fp          = file( _CSVFILE_ );
+			$device[ 0 ] = isset( $device->id ) && !empty( $device->id ) ? $device->id : count( $fp ) + 1;
+			$device[ 1 ] = implode( "|", isset( $device->names ) && !empty( $device->names ) ? $device->names : [] );
+			$device[ 2 ] = isset( $device->ip ) && !empty( $device->ip ) ? $device->ip : "";
+			$device[ 3 ] = isset( $device->username ) && !empty( $device->username ) ? $device->username : "";
+			$device[ 4 ] = isset( $device->password ) && !empty( $device->password ) ? $device->password : "";
+			$device[ 5 ] = isset( $device->img ) && !empty( $device->img ) ? $device->img : "";
+			$device[ 6 ] = isset( $device->position ) && !empty( $device->position ) ? $device->position : "";
+			
+			
+			$handle = fopen( _CSVFILE_, "a" );
+			fputcsv( $handle, $device );
+			fclose( $handle );
+		}
 		
 		/**
 		 * @param $ip